Bugfix: __main__

+++ ground_truth/extract.py
@@ -1,6 +1,6 @@
 import os; import yaml; import argparse
 
-from kipo.utils import load_data, drop_cols, encode_target, preprocessing
+from kipo.utils import load_data, extract_target_var, encode_target, preprocessing
 from utils import compute_ground_truth
 
 
@@ -17,7 +17,7 @@ def main():
     cat_cols = conf.get("cat_cols", []); num_cols = conf.get("num_cols", [])
 
     data = load_data(os.path.join(os.path.dirname(__file__), "..", "datasets", "data", DATASET), conf)
-    X, y = drop_cols(data, conf)
+    X, y = extract_target_var(data, conf)
 
     X_scaled, _ = preprocessing(X, conf)
     y_enc = encode_target(y, conf)
@@ -28,9 +28,11 @@ def main():
     print(ground_truth)
 
     os.makedirs("data", exist_ok = True); path = os.path.join("ground_truth", DATASET)
-    ground_truth.to_csv(path, index = False)
-    print(f"[INFO] ground truth rankings saved at: {path}")
+
+    if not os.path.exists(path):
+        ground_truth.to_csv(path, index = False)
+        print(f"[INFO] ground truth rankings saved at: {path}")
 
 
-if __name__ == "main":
+if __name__ == "__main__":
     main()
